retrival sys (cobert)/pyproject.toml
[build-system]
requires = ["setuptools>=62", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "retrieval-rerank-codebert"
version = "0.1.0"
description = "Retrieval and reranking system using CodeBERT and Supabase pgvector"
authors = [{ name = "you" }]
requires-python = ">=3.9"
dependencies = [
	"torch>=2.1",
	"transformers>=4.41",
	"tokenizers>=0.15",
	"sentencepiece>=0.1.99",
	"tqdm>=4.66",
	"numpy>=1.26",
	"pandas>=2.2",
	"python-dotenv>=1.0",
	"supabase>=2.5",
	"pgvector>=0.2.5",
	"psycopg2-binary>=2.9",
	"sentence-transformers>=2.7",
]

[tool.setuptools]
packages = ["src"]
